Desktop Updater (#61)
* ico file

* s3 credentials

* icons dir

* prod update url

* 1.2.0

* 1.3.0

* disable windows build

* non-blocking updater; wip

* binary install as ref

* download with resource

* restart in place

* 1.2.0

* with logs

* log error

* publish .dmg

* .dmg with /applications

* icon; drop windows build

* updater toast

* desktop logger

* Code TLC

* only render updater once

* Code TLC

+++ src/hooks/miner/use_miner_native.rs
@@ -135,7 +135,7 @@ async fn find_hash_par(
                             break;
                         } else if core_id.id == 0 {
                             let remaining = cutoff_time.saturating_sub(timer.elapsed().as_secs());
-                            solutions_channel.send(OutputMessage::TimeRemaining(remaining));
+                            solutions_channel.send(OutputMessage::TimeRemaining(remaining as i64));
                             log::info!(
                                 "Mining... Time remaining: {} {}",
                                 format_duration(remaining as u32),
